services/evaluation-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
aiokafka==0.10.0
openai==1.12.0
anthropic>=0.40.0
sqlalchemy==2.0.0
asyncpg==0.29.0
pydantic==2.5.0
pydantic-settings==2.1.0
prometheus-client==0.19.0
python-dotenv==1.0.0
pytest==7.4.0
